requirements.txt
attrs==22.2.0
cfgv==3.3.1
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.10.7
identify==2.5.22
iniconfig==2.0.0
mysql-connector-python==8.0.32
nodeenv==1.7.0
packaging==23.0
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.1
protobuf==3.20.3
pytest==7.2.2
PyYAML==6.0
tomli==2.0.1
virtualenv==20.21.0
